ru edit
I try to keep close to original, yet have slightest literary editing (when it's possible and not disturbing your idea), because without it Russian version looks a bit numb (in a bad way).

In most cases Russian proper nouns have only first letter capitalised, so 'Тысяча комнат', 'Летучая мышь', etc. 

'Беллум' is better than 'Белум'. 'Под редакцией Д. Л. Линвеги' because declension. 

English names of the characters are quite general in terms of age, sex, etc. In Russian, basic form for these entities would be Кошка, Сова, Летучая мышь, Лиса: all female and rather adult. In the context of your game it sounds weird, to my ear. I thought of different options for solving this case (and had tested quite a lot of them, trying to keep their names declaratory/abstract, avoid that gender mess, bypass all connotations which are irrelevant to the body of the story, but keep the feel and spirit of it). I think that from that perspective the best way will be to name them as Кошка, Сова, Летучая мышь и Лисёнок. Technically, Лисёнок is a 'small male fox', but in the overall context of the game this makes sense because it will balance other characters nicely, keeping everything simple but adding harmony to their crew, as well as some children book touch. It's a bit compromise which helps to avoid diminutive mess (Котёнок, Совушка, Летучая Мышка, Лисичка—ewww!). Lastly, it's probably super close to what you'd like to express with your 'Фокс'. Anyway, I'll update this part if I'll come up with the even better idea, but now it looks fine to me.

Your desk is the main question for me. Plain writing desks with openable lids are not widely common in Russia, but a generation (or half) ago these were used in a schools everywhere. Classroom desk with this lid is called 'парта' (https://ru.wikipedia.org/wiki/%D0%9F%D0%B0%D1%80%D1%82%D0%B0) and it was very natural reaction for me to think of it when I saw Recca's picture for the first time. I'm not sure if modern kids know something about парта today, thus I can call it 'письменный столик' (little writing desk—which is perfectly fine and was my first word of choice), but I would like to try use парта instead, especially considering that you name your chapters as Lessons. It looks a bit strange in the intro (why on Earth I would wake up in a room with school desk and UNBREAKABLE WALLS) so if you are not exaggerating on purpose, should I rather put it into 'вы проснулись в белой замкнутой комнате' (you woke up in the white isolated/enclosed room)? Which sounds a bit more natural in Russian, than 'Вы проснулись в комнате, окруженные белыми нерушимыми стенами'. 

Lesson 2. 'Летучая мышь находит ещё карандаш': I put 'another pencil' instead of 'second pencil' because it feels odd here being too specific. Same for Fox in Lesson 5. ('У Лисёнка еще один гость').

Lesson 3. I'm not sure if the notes which Owl finds in the desk in are the notes bearing your handwriting from the intro, so I put it unspecific ('записи'). 

Lesson 4. Does Cat literally enters the desk? Like creeps under the lid? Then it will be 'Кошка залезает в парту' (climbs inside the desk). 

What exactly Owl tries to understand? Does it literally 'understands the room', like 'solve' it? It still sounds strange at the moment so it would be better for me to have extra context here to be more accurate. 

Lesson 6. Word 'каракули' is nice, but I feel that it rather describes only graphics of a scribble and not an actual item (piece of paper). So I use 'почеркушки' for it. It's not very well documented word, but quite common in spoken language. It can refer specifically to pieces of paper filled with scribbles (usually small kid's attempts to draw; or pages filled with scribbles which someone could leave on table after having a phone call).  

Lesson 6. I would like to say that Bat finds a pencil AGAIN ('Летучая мышь СНОВА нахоидт карандаш') but not sure how it fit, so keep it simple.

I would like to read everything out after implementation to proof if everything fits perfectly for the actual game.